+++ packages/roosterjs-editor-plugins/lib/Paste/Paste.ts
@@ -156,7 +156,7 @@ export default class Paste implements EditorPlugin {
                     true /*markSelection*/
                 );
             } else {
-                this.editor.setContent(clipboardData.snapshotBeforePaste, undefined);
+                this.editor.setContent(clipboardData.snapshotBeforePaste);
             }
 
             switch (pasteOption) {
