--- packages/roosterjs-editor-types/lib/consts/DarkMode.ts
@@ -1,10 +1,0 @@
-export const DARK_MODE_DEFAULT_FORMAT = {
-    backgroundColors: {
-        darkModeColor: 'rgb(51,51,51)',
-        lightModeColor: 'rgb(255,255,255)',
-    },
-    textColors: {
-        darkModeColor: 'rgb(255,255,255)',
-        lightModeColor: 'rgb(0,0,0)',
-    }
-}
